Remove index/models.go
